Add bash scripts for mergepr and promote
Prior to this commit, we did not have scripts for signing pr merge
commits and promoting a devel branch to master.  The lack of these
scripts meant that these two tasks were happening in the web UI without
signatures.  The lack of signatures is in violation of our web of trust.
These scripts have been created so that a user need only run one
command on the command line to create a signed PR merge or branch
promotion, in keeping with our web of trust.