Refactor replaces_work_in_progress_test
Prior to this commit, one of the replaces work in progress tests
was checking md5sums to check that the file was not changed.  However,
a merge might result in changes being pulled down into the file, and
if that were the case then the test would give a false negative.  This
refactor changes the test to just check that the original content is
still in the file.