Avoid mutating local copy of feature branch on merge
Prior to this commit, if somebody attempted to merge a feature branch into
an upstream repo from their local repository using git-mergepr, and they
had a local version of the branch to be merged that differed from the remote
version, their local branch would be mutated. This commit checks out the branch
top be mergeed in a detached state, so that the process leaves the developer's
local copy of the feature branch untouched.