Add help text, scan for flags
Prior to this commit, there wasn't any help text around the mergepr or promote tools. This commit adds help text. It also converts the argument parsing of mergepr to use a read-then-shift methodology, allowing the --no-parse flag to appear anywhere, and ensuring that the correct amount of arguments are supplied to the tool.