Remove --verify-signatures from mergepr tool
Prior to this commit, the mergepr tool was passing the --verify-signatures flag to check for gpg signed commits. The --verify-signatures flag will actually fail if the gpg key exists but is not trusted. Since the current gpg keys of PolySync are not being registered as globally trusted, commits by untrusted dvelopers would fail to merge. This commit removes that flag until further decisions around this process can be made.