Revert to original state on merge failure
Prior to this commit, when the mergepr command failed, the repo might not end
up in the original state. This commit alters the failure behaviors to checkout
the original working branch and unstash changes before exiting. This commit
also adds an extra step of clearing the index before switching branches, because
in the case of a merge conflict there will be conflicts in the index that would
cause the checkout to fail, exiting with a misleading error code. If the command
fails on stash, not extra measures are taken. If it fails on fetch or checkout,
we simply pop the stash. If it fails on merge, we do a hard reset, check out the
original branch, and then pop the stash. If it fails at a later point, we do
the same thing.

fixup to  mergepr