Revert to original state on git-promote failure
Prior to this commit, if the git-promote command failed, it would exit leaving
the repo in an undesired state different from prior to running the command.
This commit ensures that on failure, the original branch is checked out and any
unsaved changes popped from the stash. If the command fails on stash, no measures
are taken. If it fails on fetch, we simply pop the stash. If it fails on merge,
we do a hard reset to clear unstaged change, checkout the original branch and then
pop the stash. If it fails at any later step we do the same.