Cleanup code
Prior to this commit, there was a lot of duplicated code and unnecessary stored variables. This commit moves all of the exit-with-error functionality into a function that prints an error message and returns a value. This value is 1 if no other value is supplied, or a specific value if one is. It also changes variable names to lowercase to prevent any accidental conflicts with shell or environment variables.