Add tests for unsigned commits and tags
Prior to this commit, we were not testing explicitly unsigned commits and tags with the mergepr and promote tools. This commit adds tests that ensure correct functionality for unsigned commits and tags. Until we are able to automatically import gpg keys we want to allow unsigned commits to be merged into devel or promoted to master.