Print helpful progress statements to stdout
Prior to this commit, the mergepr command spat out the commands it
was running and printed any stdout. In order to aid in parsing these
results, this commit adds more context surrounding which step in the
process we are in. Significantly, it lets the user know if the merge
was successful, which prevents them from being misled by later errors.