Stash untracked changes before switching branches
Prior to this commit, the git-promote and git-mergepr commands
would only stash unsaved changes in the index before attempting to
do their work. This would cause a problem when attempting to merge
a PR that mutates a file that already exists but is untracked by
the current branch. This commit stashes untracked changes as well
to avoid the potential git snafu.