reduce noisiness
Prior to this commit git-mergepr and git-promote were outputting both a
summary of the step about to complete and the command itself (and in
some cases making it harder to parse by adding RUNNING COMMAND and
OUTPUT). This commit removes the individually executed commands from the
output and only displays their output if an error occurs and, as a
compromise, no longer suppresses the normal output from each of the git
commands.