Only prune feature branch if it exists locally
Prior to this commit, if there was no local copy of the feature branch,
the git-mergepr command would fail on the auto-prune step, as it would
attempt to delete a branch that does not exist, causing the command to
complain and exit with an error, even though the merge+push step succeeded.
This commit alters the script to only attempt to delete the branch if it does
exist. However, if the local branch contains commits that have not been
fully merged, mergepr will still attempt to delete it and fail.