Add tests for running on feature branch without --no-prune
Prior to this commit, the behavior of mergepr while on the feature branch
to be merged without passing the --no-prune flag was unspecified. This commit
adds tests that specify this behavior. If the command is run as described, the
script should exit with an error, suggesting that the user either run with the
--no-prune flag if they want to keep the branch or else run from a different
branch.