Update tests with new expected behavior
Prior to this commit, a few tests would fail because they set up
the scenario for merge-pr tests on the branch to be merged. Since
the tool has been rewritten to fail when run on the branch to be
merged with the default behavior, these tests were failing. This commit
alters these tests to begin the process from another branch. One step whose
purpose is to create a merge conflict was leaving the repo checked out on
the feaature branch. Now it checks out the branch it was on at the beginning
of the step. Another scenario explicitly tested the behavior that has since
been altered. Where we used to expect failure on re-checkout (8), now we expect
failure before we even begin (10). The scenario is updated accordingly.