Add argument flexibility to git-secure-push
Prior to this commit, git-securepush was only allowed two inputs. It could be passed a
branch name, in which case it would try to push that branch to its remote counterpart,
assuming the repo was currently checked out on that branch. Its default
behavior, when passed no arguments was to push the current branch. In order to use
secure-push in the git-merge-pr command, it needs to allow pushing from a
detached HEAD state. This commit alters secure-push to allow different kinds of
inputs. It now allows the following:

git secure-push
git secure-push HEAD
git-secure-push HEAD:branch_name
git-secure-push branch_name:branch_name

with the expected behavior for each.