Generate push entries from dev/urandom
Prior to this commit, the RSL push entries were created using `dev/random`.
When there was not enough entropy in the system to generate 16 pseudorandom
bytes, `dev/random` would try to block until it could create more. However,
the `dd` command being used to get the bytes was not waiting, and would exit
with 12, 13, 15 or however many bytes got generated. In order to consistently
generate 16 byte strings, this commit switches to use `dev/urandom`, which
reuses the entropy pool numbers to create unlimited pseudo-random bits.