Alter git-merge-pr to use secure-push and secure-fetch
Prior to this commit, git-merge-pr used regular fetch and push internally.
This commit updates them to use secure versions of these commands, so the
RSL is updated during the operation. Since secure-fetch can't be called
with the `--prune` flag, the git cleanup command is invoked in its place
to do local housekeeping.