Use secure-fetch and secure-push inside git-promote
Prior to this commit, git-promote used regular old insecure fetch and push.
This commit updates it to use fancy customized secure-fetch and secure-push, in
order to make promote secure. One of the tests had to be updated because if the
GPG signing key is unavailable, promote will fail before the sign step. It will
fail on the secure-fetch step, which requires a valid gpg key as well.