Initial version of various scripts
Prior to this commit, we did not have any scripts to aid in our workflow
for merging pull requests with GPG-signed commits, or promoting signed
pre-release tags to signed release tags, or mitigating metadata
manipulation attacks. This commit adds

* `git-merge-pr`
* `git-promote`
* `git-secure-fetch`
* `git-secure-pull`
* `git-secure-push`

for handling those various scenarios, in addition to an `install.sh` for
adding these git helpers to `~/.local/bin` and ensuring that's part of
the user's `$PATH`.