Copy commands by default
Prior to this commit, the install script created symlinks to a directory in the
user's path. This commit changes the default behavior to copy the commands to
the target directory. In order to create symlinks instead, the user can pass the
`--dev` option to `install.sh`. This option facilitates installation for those
who desire to use, but not edit the secure git commands. After installation, as
long as the files are copied, the user can safely delete the repo. To update,
the user can re-clone and re-install.