Add git overrides
Prior to this commit, users had to remember to use the secure-push, secure-fetch, and secure-pull commands when using git. This commit adds a part of the install script to set an alias in ~/.bashrc, enabling kevlar-laces to intercept git commands and prepend them with `secure-` if they are push, pull, or fetch, preventing accidental unsecured actions. This updates the README to reflect those changes. This functionality is optional, and can be disabled by passing the `--no-override` flag to the install script.