Fetch upstream revisions
Prior to this commit, merge-pr fetched the latest revisions for whichever branch
they happened to be on when they ran the command. If the user did not already
have up to date copies of both the target branch and the branch to be merged
into it, they ran the risk of pushing incorrect commit history to the remote git
server. This commit securely fetches revisions of both branches before
attemoting to do any merging.