Run commands from any directory inside project
Prior to this commit, if any of the secure git commands were run outside of the
top-level git directory, they would enter into an endless loop looking for files
that exist at the root level. This commit offers a temporary fix. It alters the
commands to operate at the project root before returning to the starting
directory after work is complete.